Refactor server to improvement arch
